[#2] Refactor : 컴포넌트 분리 및 코드 정리
- useGetData hook을 만들어, 데이터를 가져오고 가공하는 부분을 따로 빼서 관리
- tsconfig.json에서 target을 es6로 바꾸어, new Set 문법 적용
